+++ yash-syntax/src/parser/pipeline.rs
@@ -105,7 +105,7 @@ mod tests {
     use super::super::error::ErrorCause;
     use super::super::lex::Lexer;
     use super::*;
-    use crate::alias::{AliasSet, HashEntry};
+    use crate::alias::{AliasSet, EmptyGlossary, HashEntry};
     use crate::source::Location;
     use crate::source::Source;
     use futures_util::FutureExt;
@@ -113,8 +113,7 @@ mod tests {
     #[test]
     fn parser_pipeline_eof() {
         let mut lexer = Lexer::from_memory("", Source::Unknown);
-        let aliases = Default::default();
-        let mut parser = Parser::new(&mut lexer, &aliases);
+        let mut parser = Parser::new(&mut lexer, &EmptyGlossary);
 
         let option = parser.pipeline().now_or_never().unwrap().unwrap().unwrap();
         assert_eq!(option, None);
@@ -123,8 +122,7 @@ mod tests {
     #[test]
     fn parser_pipeline_one() {
         let mut lexer = Lexer::from_memory("foo", Source::Unknown);
-        let aliases = Default::default();
-        let mut parser = Parser::new(&mut lexer, &aliases);
+        let mut parser = Parser::new(&mut lexer, &EmptyGlossary);
 
         let result = parser.pipeline().now_or_never().unwrap();
         let p = result.unwrap().unwrap().unwrap();
@@ -136,8 +134,7 @@ mod tests {
     #[test]
     fn parser_pipeline_many() {
         let mut lexer = Lexer::from_memory("one | two | \n\t\n three", Source::Unknown);
-        let aliases = Default::default();
-        let mut parser = Parser::new(&mut lexer, &aliases);
+        let mut parser = Parser::new(&mut lexer, &EmptyGlossary);
 
         let result = parser.pipeline().now_or_never().unwrap();
         let p = result.unwrap().unwrap().unwrap();
@@ -151,8 +148,7 @@ mod tests {
     #[test]
     fn parser_pipeline_negated() {
         let mut lexer = Lexer::from_memory("! foo", Source::Unknown);
-        let aliases = Default::default();
-        let mut parser = Parser::new(&mut lexer, &aliases);
+        let mut parser = Parser::new(&mut lexer, &EmptyGlossary);
 
         let result = parser.pipeline().now_or_never().unwrap();
         let p = result.unwrap().unwrap().unwrap();
@@ -164,8 +160,7 @@ mod tests {
     #[test]
     fn parser_pipeline_double_negation() {
         let mut lexer = Lexer::from_memory(" !  !", Source::Unknown);
-        let aliases = Default::default();
-        let mut parser = Parser::new(&mut lexer, &aliases);
+        let mut parser = Parser::new(&mut lexer, &EmptyGlossary);
 
         let e = parser.pipeline().now_or_never().unwrap().unwrap_err();
         assert_eq!(e.cause, ErrorCause::Syntax(SyntaxError::DoubleNegation));
@@ -178,8 +173,7 @@ mod tests {
     #[test]
     fn parser_pipeline_missing_command_after_negation() {
         let mut lexer = Lexer::from_memory("!\nfoo", Source::Unknown);
-        let aliases = Default::default();
-        let mut parser = Parser::new(&mut lexer, &aliases);
+        let mut parser = Parser::new(&mut lexer, &EmptyGlossary);
 
         let e = parser.pipeline().now_or_never().unwrap().unwrap_err();
         assert_eq!(
@@ -195,8 +189,7 @@ mod tests {
     #[test]
     fn parser_pipeline_missing_command_after_bar() {
         let mut lexer = Lexer::from_memory("foo | ;", Source::Unknown);
-        let aliases = Default::default();
-        let mut parser = Parser::new(&mut lexer, &aliases);
+        let mut parser = Parser::new(&mut lexer, &EmptyGlossary);
 
         let e = parser.pipeline().now_or_never().unwrap().unwrap_err();
         assert_eq!(
@@ -212,8 +205,7 @@ mod tests {
     #[test]
     fn parser_pipeline_bang_after_bar() {
         let mut lexer = Lexer::from_memory("foo | !", Source::Unknown);
-        let aliases = Default::default();
-        let mut parser = Parser::new(&mut lexer, &aliases);
+        let mut parser = Parser::new(&mut lexer, &EmptyGlossary);
 
         let e = parser.pipeline().now_or_never().unwrap().unwrap_err();
         assert_eq!(e.cause, ErrorCause::Syntax(SyntaxError::BangAfterBar));